move location of auth service
