+++ src/components/AuthProvider.tsx
@@ -3,7 +3,7 @@
 import { useEffect } from "react";
 import { clientSignIn } from "@/store/features/auth/authSlice";
 import { useAppDispatch } from "@/store/hooks";
-import { getUser } from "@/api/authService";
+import { getUser } from "@/app/(auth)/authService";
 
 export default function AuthProvider() {
   const dispatch = useAppDispatch();
